refined hostmode networking
- skip disableTxOffload for hostmode containers
- verification that hostmode containers do not have additional links defined in topology
- swapped linux.go and vr-xrv.go content, content was wrongly associated to files
