+++ clab/linux.go
@@ -1,31 +1,26 @@
 package clab
 
-import "fmt"
+import "strings"
 
-func initVrXRVNode(c *CLab, nodeCfg NodeConfig, node *Node, user string, envs map[string]string) error {
+func initLinuxNode(c *CLab, nodeCfg NodeConfig, node *Node, user string, envs map[string]string) error {
 	var err error
 
+	node.Config, err = c.configInit(&nodeCfg, node.Kind)
+	if err != nil {
+		return err
+	}
 	node.Image = c.imageInitialization(&nodeCfg, node.Kind)
 	node.Group = c.groupInitialization(&nodeCfg, node.Kind)
 	node.Position = c.positionInitialization(&nodeCfg, node.Kind)
+	node.Cmd = c.cmdInit(&nodeCfg, node.Kind)
 	node.User = user
 
-	// env vars are used to set launch.py arguments in vrnetlab container
-	defEnv := map[string]string{
-		"USERNAME":           "clab",
-		"PASSWORD":           "clab@123",
-		"CONNECTION_MODE":    vrDefConnMode,
-		"DOCKER_NET_V4_ADDR": c.Config.Mgmt.IPv4Subnet,
-		"DOCKER_NET_V6_ADDR": c.Config.Mgmt.IPv6Subnet,
-	}
-	node.Env = mergeStringMaps(defEnv, envs)
-
-	if node.Env["CONNECTION_MODE"] == "macvtap" {
-		// mount dev dir to enable macvtap
-		node.Binds = append(node.Binds, "/dev:/dev")
+	node.Sysctls = make(map[string]string)
+	if strings.ToLower(node.NetworkMode) != "host" {
+		node.Sysctls["net.ipv6.conf.all.disable_ipv6"] = "0"
 	}
 
-	node.Cmd = fmt.Sprintf("--username %s --password %s --hostname %s --connection-mode %s --trace", node.Env["USERNAME"], node.Env["PASSWORD"], node.ShortName, node.Env["CONNECTION_MODE"])
+	node.Env = envs
 
 	return err
 }
