+++ clab/vr-xrv.go
@@ -1,22 +1,31 @@
 package clab
 
-func initLinuxNode(c *CLab, nodeCfg NodeConfig, node *Node, user string, envs map[string]string) error {
+import "fmt"
+
+func initVrXRVNode(c *CLab, nodeCfg NodeConfig, node *Node, user string, envs map[string]string) error {
 	var err error
 
-	node.Config, err = c.configInit(&nodeCfg, node.Kind)
-	if err != nil {
-		return err
-	}
 	node.Image = c.imageInitialization(&nodeCfg, node.Kind)
 	node.Group = c.groupInitialization(&nodeCfg, node.Kind)
 	node.Position = c.positionInitialization(&nodeCfg, node.Kind)
-	node.Cmd = c.cmdInit(&nodeCfg, node.Kind)
 	node.User = user
 
-	node.Sysctls = make(map[string]string)
-	node.Sysctls["net.ipv6.conf.all.disable_ipv6"] = "0"
+	// env vars are used to set launch.py arguments in vrnetlab container
+	defEnv := map[string]string{
+		"USERNAME":           "clab",
+		"PASSWORD":           "clab@123",
+		"CONNECTION_MODE":    vrDefConnMode,
+		"DOCKER_NET_V4_ADDR": c.Config.Mgmt.IPv4Subnet,
+		"DOCKER_NET_V6_ADDR": c.Config.Mgmt.IPv6Subnet,
+	}
+	node.Env = mergeStringMaps(defEnv, envs)
 
-	node.Env = envs
+	if node.Env["CONNECTION_MODE"] == "macvtap" {
+		// mount dev dir to enable macvtap
+		node.Binds = append(node.Binds, "/dev:/dev")
+	}
+
+	node.Cmd = fmt.Sprintf("--username %s --password %s --hostname %s --connection-mode %s --trace", node.Env["USERNAME"], node.Env["PASSWORD"], node.ShortName, node.Env["CONNECTION_MODE"])
 
 	return err
 }
